discord_bot/requirements.txt
discord.py==2.1.0
feedparser>=6.0<=7
html2text==2020.1.16
python-dateutil>=2.8.0<=3
pytz==2022.1
requests>=2.27.1<=3
pyyaml==6.0
pydash==5.1.0
tweepy==4.10.1
praw==7.6.0
twitchAPI==2.5.7